+++ Projects/Server/Client/UOPFiles.cs
@@ -13,6 +13,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
  *************************************************************************/
 
+using System;
 using System.Collections.Generic;
 using System.IO;
 
@@ -20,7 +21,9 @@ namespace Server;
 
 public static class UOPFiles
 {
-    public static Dictionary<int, UOPEntry> ReadUOPIndexes(FileStream stream, string fileExt, int entryCount = 0x14000, int expectedVersion = -1)
+    public static Dictionary<int, UOPEntry> ReadUOPIndexes(
+        FileStream stream, string fileExt, int entryCount = 0x14000, int expectedVersion = -1
+    )
     {
         var reader = new BinaryReader(stream);
         var length = (int)stream.Length;
@@ -92,9 +95,141 @@ public static class UOPFiles
 
         for (var i = 0; i < entryCount; i++)
         {
-            hashes[UOPHash.HashLittle2($"{root}/{i:D8}{ext}")] = i;
+            hashes[HashLittle2($"{root}/{i:D8}{ext}")] = i;
         }
 
         return hashes;
+    }
+
+    public static ulong HashLittle2(ReadOnlySpan<char> s)
+    {
+        var length = s.Length;
+
+        uint b, c;
+        var a = b = c = 0xDEADBEEF + (uint)length;
+
+        var k = 0;
+
+        while (length > 12)
+        {
+            a += s[k];
+            a += (uint)s[k + 1] << 8;
+            a += (uint)s[k + 2] << 16;
+            a += (uint)s[k + 3] << 24;
+            b += s[k + 4];
+            b += (uint)s[k + 5] << 8;
+            b += (uint)s[k + 6] << 16;
+            b += (uint)s[k + 7] << 24;
+            c += s[k + 8];
+            c += (uint)s[k + 9] << 8;
+            c += (uint)s[k + 10] << 16;
+            c += (uint)s[k + 11] << 24;
+
+            a -= c;
+            a ^= (c << 4) | (c >> 28);
+            c += b;
+            b -= a;
+            b ^= (a << 6) | (a >> 26);
+            a += c;
+            c -= b;
+            c ^= (b << 8) | (b >> 24);
+            b += a;
+            a -= c;
+            a ^= (c << 16) | (c >> 16);
+            c += b;
+            b -= a;
+            b ^= (a << 19) | (a >> 13);
+            a += c;
+            c -= b;
+            c ^= (b << 4) | (b >> 28);
+            b += a;
+
+            length -= 12;
+            k += 12;
+        }
+
+        if (length != 0)
+        {
+            switch (length)
+            {
+                case 12:
+                    {
+                        c += (uint)s[k + 11] << 24;
+                        goto case 11;
+                    }
+                case 11:
+                    {
+                        c += (uint)s[k + 10] << 16;
+                        goto case 10;
+                    }
+                case 10:
+                    {
+                        c += (uint)s[k + 9] << 8;
+                        goto case 9;
+                    }
+                case 9:
+                    {
+                        c += s[k + 8];
+                        goto case 8;
+                    }
+                case 8:
+                    {
+                        b += (uint)s[k + 7] << 24;
+                        goto case 7;
+                    }
+                case 7:
+                    {
+                        b += (uint)s[k + 6] << 16;
+                        goto case 6;
+                    }
+                case 6:
+                    {
+                        b += (uint)s[k + 5] << 8;
+                        goto case 5;
+                    }
+                case 5:
+                    {
+                        b += s[k + 4];
+                        goto case 4;
+                    }
+                case 4:
+                    {
+                        a += (uint)s[k + 3] << 24;
+                        goto case 3;
+                    }
+                case 3:
+                    {
+                        a += (uint)s[k + 2] << 16;
+                        goto case 2;
+                    }
+                case 2:
+                    {
+                        a += (uint)s[k + 1] << 8;
+                        goto case 1;
+                    }
+                case 1:
+                    {
+                        a += s[k];
+                        break;
+                    }
+            }
+
+            c ^= b;
+            c -= (b << 14) | (b >> 18);
+            a ^= c;
+            a -= (c << 11) | (c >> 21);
+            b ^= a;
+            b -= (a << 25) | (a >> 7);
+            c ^= b;
+            c -= (b << 16) | (b >> 16);
+            a ^= c;
+            a -= (c << 4) | (c >> 28);
+            b ^= a;
+            b -= (a << 14) | (a >> 18);
+            c ^= b;
+            c -= (b << 24) | (b >> 8);
+        }
+
+        return ((ulong)b << 32) | c;
     }
 }
