+++ Projects/Server/MultiData.cs
@@ -736,150 +736,3 @@ public sealed class MultiComponentList
         }
     }
 }
-
-public static class UOPHash
-{
-    public static void BuildChunkIDs(out Dictionary<ulong, int> chunkIds)
-    {
-        const int maxId = 0x10000;
-
-        chunkIds = new Dictionary<ulong, int>();
-
-        for (var i = 0; i < maxId; ++i)
-        {
-            chunkIds[HashLittle2($"build/multicollection/{i:000000}.bin")] = i;
-        }
-    }
-
-    public static ulong HashLittle2(ReadOnlySpan<char> s)
-    {
-        var length = s.Length;
-
-        uint b, c;
-        var a = b = c = 0xDEADBEEF + (uint)length;
-
-        var k = 0;
-
-        while (length > 12)
-        {
-            a += s[k];
-            a += (uint)s[k + 1] << 8;
-            a += (uint)s[k + 2] << 16;
-            a += (uint)s[k + 3] << 24;
-            b += s[k + 4];
-            b += (uint)s[k + 5] << 8;
-            b += (uint)s[k + 6] << 16;
-            b += (uint)s[k + 7] << 24;
-            c += s[k + 8];
-            c += (uint)s[k + 9] << 8;
-            c += (uint)s[k + 10] << 16;
-            c += (uint)s[k + 11] << 24;
-
-            a -= c;
-            a ^= (c << 4) | (c >> 28);
-            c += b;
-            b -= a;
-            b ^= (a << 6) | (a >> 26);
-            a += c;
-            c -= b;
-            c ^= (b << 8) | (b >> 24);
-            b += a;
-            a -= c;
-            a ^= (c << 16) | (c >> 16);
-            c += b;
-            b -= a;
-            b ^= (a << 19) | (a >> 13);
-            a += c;
-            c -= b;
-            c ^= (b << 4) | (b >> 28);
-            b += a;
-
-            length -= 12;
-            k += 12;
-        }
-
-        if (length != 0)
-        {
-            switch (length)
-            {
-                case 12:
-                    {
-                        c += (uint)s[k + 11] << 24;
-                        goto case 11;
-                    }
-                case 11:
-                    {
-                        c += (uint)s[k + 10] << 16;
-                        goto case 10;
-                    }
-                case 10:
-                    {
-                        c += (uint)s[k + 9] << 8;
-                        goto case 9;
-                    }
-                case 9:
-                    {
-                        c += s[k + 8];
-                        goto case 8;
-                    }
-                case 8:
-                    {
-                        b += (uint)s[k + 7] << 24;
-                        goto case 7;
-                    }
-                case 7:
-                    {
-                        b += (uint)s[k + 6] << 16;
-                        goto case 6;
-                    }
-                case 6:
-                    {
-                        b += (uint)s[k + 5] << 8;
-                        goto case 5;
-                    }
-                case 5:
-                    {
-                        b += s[k + 4];
-                        goto case 4;
-                    }
-                case 4:
-                    {
-                        a += (uint)s[k + 3] << 24;
-                        goto case 3;
-                    }
-                case 3:
-                    {
-                        a += (uint)s[k + 2] << 16;
-                        goto case 2;
-                    }
-                case 2:
-                    {
-                        a += (uint)s[k + 1] << 8;
-                        goto case 1;
-                    }
-                case 1:
-                    {
-                        a += s[k];
-                        break;
-                    }
-            }
-
-            c ^= b;
-            c -= (b << 14) | (b >> 18);
-            a ^= c;
-            a -= (c << 11) | (c >> 21);
-            b ^= a;
-            b -= (a << 25) | (a >> 7);
-            c ^= b;
-            c -= (b << 16) | (b >> 16);
-            a ^= c;
-            a -= (c << 4) | (c >> 28);
-            b ^= a;
-            b -= (a << 14) | (a >> 18);
-            c ^= b;
-            c -= (b << 24) | (b >> 8);
-        }
-
-        return ((ulong)b << 32) | c;
-    }
-}
